feat: 트랜잭션 추가 및 기본데이터 추가
