+++ src/main/java/in/koreatech/koin/domain/shop/scheduler/NotificationScheduler.java
@@ -1,24 +1,9 @@
 package in.koreatech.koin.domain.shop.scheduler;
 
-import static in.koreatech.koin.global.fcm.MobileAppPath.*;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import in.koreatech.koin.domain.shop.model.shop.Shop;
-import in.koreatech.koin.domain.shop.model.shop.ShopNotificationMessage;
-import in.koreatech.koin.domain.shop.model.shop.ShopNotificationQueue;
-import in.koreatech.koin.domain.shop.repository.shop.ShopCategoryMapRepository;
-import in.koreatech.koin.domain.shop.repository.shop.ShopNotificationQueueRepository;
-import in.koreatech.koin.domain.shop.repository.shop.ShopRepository;
-import in.koreatech.koin.domain.user.model.User;
-import in.koreatech.koin.domain.user.repository.UserRepository;
-import in.koreatech.koin.global.domain.notification.model.Notification;
-import in.koreatech.koin.global.domain.notification.model.NotificationFactory;
-import in.koreatech.koin.global.domain.notification.service.NotificationService;
+import in.koreatech.koin.domain.shop.service.NotificationScheduleService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -27,45 +12,14 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class NotificationScheduler {
 
-    private final ShopNotificationQueueRepository shopNotificationQueueRepository;
-    private final NotificationFactory notificationFactory;
-    private final NotificationService notificationService;
-    private final ShopRepository shopRepository;
-    private final ShopCategoryMapRepository shopCategoryMapRepository;
-    private final UserRepository userRepository;
+    private final NotificationScheduleService notificationScheduleService;
 
     @Scheduled(cron = "0 * * * * *")
     public void sendDueNotifications() {
         try {
-            LocalDateTime now = LocalDateTime.now();
-
-            List<Notification> notifications = shopNotificationQueueRepository
-                .findByNotificationTimeBefore(now)
-                .stream()
-                .map(this::createNotification)
-                .toList();
-
-            shopNotificationQueueRepository.deleteByNotificationTimeBefore(now);
-
-            notificationService.push(notifications);
+            notificationScheduleService.sendDueNotifications();
         } catch (Exception e) {
             log.warn("리뷰유도 알림 전송 과정에서 오류가 발생했습니다.");
         }
-    }
-
-    private Notification createNotification(ShopNotificationQueue dueNotification) {
-        Shop shop = shopRepository.getById(dueNotification.getShopId());
-        ShopNotificationMessage shopNotificationMessage =
-            shopCategoryMapRepository.findNotificationMessageByShopId(dueNotification.getShopId());
-        User user = userRepository.getById(dueNotification.getUserId());
-
-        return notificationFactory.generateReviewPromptNotification(
-            SHOP,
-            dueNotification.getShopId(),
-            shop.getName(),
-            shopNotificationMessage.getTitle(),
-            shopNotificationMessage.getContent(),
-            user
-        );
     }
 }
