+++ src/test/java/in/koreatech/koin/acceptance/ShopApiTest.java
@@ -2,15 +2,23 @@ package in.koreatech.koin.acceptance;
 
 import static in.koreatech.koin.domain.shop.model.review.ReportStatus.DISMISSED;
 import static in.koreatech.koin.domain.shop.model.review.ReportStatus.UNHANDLED;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.time.Clock;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZoneOffset;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
+import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,6 +27,7 @@ import in.koreatech.koin.domain.owner.model.Owner;
 import in.koreatech.koin.domain.shop.model.menu.Menu;
 import in.koreatech.koin.domain.shop.model.review.ShopReview;
 import in.koreatech.koin.domain.shop.model.shop.Shop;
+import in.koreatech.koin.domain.shop.scheduler.NotificationScheduler;
 import in.koreatech.koin.domain.student.model.Student;
 import in.koreatech.koin.fixture.EventArticleFixture;
 import in.koreatech.koin.fixture.MenuCategoryFixture;
@@ -28,6 +37,7 @@ import in.koreatech.koin.fixture.ShopFixture;
 import in.koreatech.koin.fixture.ShopReviewFixture;
 import in.koreatech.koin.fixture.ShopReviewReportFixture;
 import in.koreatech.koin.fixture.UserFixture;
+import in.koreatech.koin.global.domain.notification.service.NotificationService;
 
 @Transactional
 @SuppressWarnings("NonAsciiCharacters")
@@ -58,10 +68,20 @@ class ShopApiTest extends AcceptanceTest {
     @Autowired
     private ShopCategoryFixture shopCategoryFixture;
 
+    @Autowired
+    private NotificationScheduler notificationScheduler;
+
+    @Autowired
+    private NotificationService notificationService;
+
+    @Mock
+    private Clock mockClock;
+
     private Shop 마슬랜;
     private Owner owner;
 
     private Student 익명_학생;
+    private String token_익명;
 
     @BeforeAll
     void setUp() {
@@ -69,6 +89,7 @@ class ShopApiTest extends AcceptanceTest {
         owner = userFixture.준영_사장님();
         마슬랜 = shopFixture.마슬랜(owner);
         익명_학생 = userFixture.익명_학생();
+        token_익명 = userFixture.getToken(익명_학생.getUser());
     }
 
     @Test
@@ -1017,4 +1038,32 @@ class ShopApiTest extends AcceptanceTest {
                 }
                 """, 티바_영업여부, 마슬랜_영업여부)));
     }
+
+    // @Test
+    // void 전화하기_이벤트_발생시_1시간_뒤에_알림을_발송한다() throws Exception {
+    //     mockMvc.perform(
+    //             post("/shops/{shopId}/call-notification", 마슬랜.getId())
+    //                 .header("Authorization", "Bearer " + token_익명)
+    //         )
+    //         .andExpect(status().isOk());
+    //     forceVerify(() -> verify(notificationEventListener, times(1))
+    //         .onNotificationEventCreate(any()));
+    //
+    //     // 이러면 DB에까진 제대로 저장된거야.
+    //     // cron 작업 테스트를 어떻게하지?
+    //
+    //     LocalDateTime oneHourTenMinutesLater = LocalDateTime.now().plusMinutes(70);
+    //     when(mockClock.instant()).thenReturn(oneHourTenMinutesLater.toInstant(ZoneOffset.UTC));
+    //     when(mockClock.getZone()).thenReturn(ZoneId.systemDefault());
+    //
+    //     notificationScheduler.sendDueNotifications();
+    //
+    //     forceVerify(() -> verify(notificationService, times(1))
+    //         .push(anyList()));
+    //
+    //     // DB에 남아있는지 확인
+    //
+    //     clear();
+    //     setUp();
+    // }
 }
